Move artist dashboard tracks table to own file
