+++ packages/web/src/pages/artist-dashboard-page/ArtistDashboardPage.tsx
@@ -1,52 +1,25 @@
-import React, {
-  Suspense,
-  Component,
-  useMemo,
-  ReactNode,
-  useCallback,
-  useState
-} from 'react'
+import React, { Suspense, Component, ReactNode } from 'react'
 
 import {
   ID,
   Status,
   Theme,
   Track,
-  User,
-  formatCurrencyBalance,
   formatCount,
   themeSelectors,
   FeatureFlags
 } from '@audius/common'
-import {
-  IconFilter,
-  IconNote,
-  IconHidden,
-  IconKebabHorizontal,
-  IconQuestionCircle,
-  HarmonyButton,
-  HarmonyButtonType,
-  PopupMenu,
-  PopupMenuItem,
-  HarmonyPlainButton,
-  HarmonyPlainButtonType
-} from '@audius/stems'
 import cn from 'classnames'
 import { push as pushRoute } from 'connected-react-router'
 import { each } from 'lodash'
 import moment, { Moment } from 'moment'
-import { connect, useDispatch, useSelector } from 'react-redux'
+import { connect } from 'react-redux'
 import { withRouter, RouteComponentProps } from 'react-router-dom'
 import { Dispatch } from 'redux'
 
-import { Icon } from 'components/Icon'
 import Header from 'components/header/desktop/Header'
-import { Input } from 'components/input'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import Page from 'components/page/Page'
-import { TracksTable, TracksTableColumn } from 'components/tracks-table'
-import { Text } from 'components/typography'
-import useTabs, { useTabRecalculator } from 'hooks/useTabs/useTabs'
 import { getFeatureEnabled } from 'services/remote-config/featureFlagHelpers'
 import { AppState } from 'store/types'
 import lazyWithPreload from 'utils/lazyWithPreload'
@@ -56,15 +29,19 @@ import { withClassNullGuard } from 'utils/withNullGuard'
 import styles from './ArtistDashboardPage.module.css'
 import ArtistProfile from './components/ArtistProfile'
 import {
+  TracksTableContainer,
+  DataSourceTrack,
+  tablePageSize
+} from './components/TracksTableContainer'
+import { USDCTile } from './components/USDCTile'
+import {
   fetchDashboard,
   fetchDashboardListenData,
-  fetchDashboardTracks,
   resetDashboard
 } from './store/actions'
 import {
   getDashboardListenData,
   getDashboardStatus,
-  getDashboardTracksStatus,
   makeGetDashboard
 } from './store/selectors'
 const { getTheme } = themeSelectors
@@ -82,259 +59,8 @@ const StatTile = (props: { title: string; value: any }) => {
   )
 }
 
-type DataSourceTrack = Track & {
-  key: string
-  name: string
-  date: string
-  time?: number
-  saves: number
-  reposts: number
-  plays: number
-}
-
-type TracksTableProps = {
-  onClickRow: (record: any) => void
-  unlistedDataSource: DataSourceTrack[]
-  listedDataSource: DataSourceTrack[]
-  account: User
-}
-
 export const messages = {
-  publicTracksTabTitle: 'PUBLIC TRACKS',
-  unlistedTracksTabTitle: 'HIDDEN TRACKS',
-  filterInputPlacehoder: 'Filter Tracks',
-  thisYear: 'This Year',
-  usdc: 'USDC',
-  earn: 'Earn USDC by selling your music',
-  learnMore: 'Learn More',
-  withdraw: 'Withdraw Funds',
-  salesSummary: 'Sales Summary',
-  withdrawalHistory: 'Withdrawal History'
-}
-
-const tableColumns: TracksTableColumn[] = [
-  'spacer',
-  'trackName',
-  'releaseDate',
-  'length',
-  'plays',
-  'reposts',
-  'overflowMenu'
-]
-
-// Pagination Constants
-const tablePageSize = 50
-
-const TracksTableContainer = ({
-  onClickRow,
-  listedDataSource,
-  unlistedDataSource,
-  account
-}: TracksTableProps) => {
-  const [filterText, setFilterText] = useState('')
-  const dispatch = useDispatch()
-  const tracksStatus = useSelector(getDashboardTracksStatus)
-  const tabRecalculator = useTabRecalculator()
-
-  const tabHeaders = useMemo(
-    () => [
-      {
-        text: messages.publicTracksTabTitle,
-        icon: <IconNote />,
-        label: messages.publicTracksTabTitle
-      },
-      {
-        text: messages.unlistedTracksTabTitle,
-        icon: <IconHidden />,
-        label: messages.unlistedTracksTabTitle,
-        disabled: !unlistedDataSource.length,
-        disabledTooltipText: 'You have no hidden tracks'
-      }
-    ],
-    [unlistedDataSource]
-  )
-
-  const handleFilterChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const val = e.target.value
-    setFilterText(val)
-  }
-
-  const filteredListedData = listedDataSource.filter((data) =>
-    data.title.toLowerCase().includes(filterText.toLowerCase())
-  )
-
-  const filteredUnlistedData = unlistedDataSource.filter((data) =>
-    data.title.toLowerCase().includes(filterText.toLowerCase())
-  )
-
-  const handleFetchPage = useCallback(
-    (page: number) => {
-      dispatch(fetchDashboardTracks(page * tablePageSize, tablePageSize))
-    },
-    [dispatch]
-  )
-
-  const tabElements = useMemo(
-    () => [
-      <div
-        key='listed'
-        className={cn(styles.sectionContainer, styles.tabBodyWrapper)}
-      >
-        <TracksTable
-          data={filteredListedData}
-          disabledTrackEdit
-          columns={tableColumns}
-          onClickRow={onClickRow}
-          loading={tracksStatus === Status.LOADING}
-          fetchPage={handleFetchPage}
-          pageSize={tablePageSize}
-          userId={account.user_id}
-          showMoreLimit={5}
-          onShowMoreToggle={tabRecalculator.recalculate}
-          totalRowCount={account.track_count}
-          isPaginated
-        />
-      </div>,
-      <div
-        key='unlisted'
-        className={cn(styles.sectionContainer, styles.tabBodyWrapper)}
-      >
-        <TracksTable
-          data={filteredUnlistedData}
-          disabledTrackEdit
-          columns={tableColumns}
-          onClickRow={onClickRow}
-          loading={tracksStatus === Status.LOADING}
-          fetchPage={handleFetchPage}
-          pageSize={tablePageSize}
-          showMoreLimit={5}
-          userId={account.user_id}
-          onShowMoreToggle={tabRecalculator.recalculate}
-          totalRowCount={account.track_count}
-          isPaginated
-        />
-      </div>
-    ],
-    [
-      account,
-      filteredListedData,
-      filteredUnlistedData,
-      handleFetchPage,
-      onClickRow,
-      tabRecalculator,
-      tracksStatus
-    ]
-  )
-
-  const { tabs, body } = useTabs({
-    bodyClassName: styles.tabBody,
-    isMobile: false,
-    tabRecalculator,
-    tabs: tabHeaders,
-    elements: tabElements
-  })
-
-  return (
-    <div className={styles.tableContainer}>
-      <div className={styles.tabBorderProvider}>
-        <div className={styles.filterInputContainer}>
-          <Input
-            placeholder={messages.filterInputPlacehoder}
-            prefix={<IconFilter />}
-            onChange={handleFilterChange}
-            value={filterText}
-            size='small'
-            variant='bordered'
-          />
-        </div>
-        <div className={styles.tabContainer}>{tabs}</div>
-      </div>
-      {body}
-    </div>
-  )
-}
-
-const USDCSection = ({ account }: { account: User }) => {
-  if (!account) return null
-
-  // TODO: wire up balance https://linear.app/audius/issue/PAY-1761/wire-up-usdc-balance-in-artist-dashboard
-  const balance = 10.29
-
-  const menuItems: PopupMenuItem[] = [
-    {
-      text: messages.salesSummary,
-      // TODO: link to sales page https://linear.app/audius/issue/PAY-1763/wire-up-salespurchases-pages-on-artist-dashboard
-      onClick: () => {}
-    },
-    {
-      text: messages.withdrawalHistory,
-      // TODO: link to withdraw history page https://linear.app/audius/issue/PAY-1763/wire-up-salespurchases-pages-on-artist-dashboard
-      onClick: () => {}
-    }
-  ]
-
-  return (
-    <div className={styles.usdcContainer}>
-      <div className={styles.backgroundBlueGradient}>
-        <div className={styles.usdcTitleContainer}>
-          <div className={styles.usdcTitle}>
-            {/* TODO: update icon https://linear.app/audius/issue/PAY-1764/update-icons-in-usdc-tile */}
-            <Icon icon={IconNote} size='xxxLarge' color='staticWhite' />
-            <div className={styles.usdc}>
-              <Text
-                variant='heading'
-                size='xxLarge'
-                color='staticWhite'
-                strength='strong'
-              >
-                {messages.usdc}
-              </Text>
-            </div>
-          </div>
-          <Text
-            variant='heading'
-            color='staticWhite'
-            strength='strong'
-            size='xxLarge'
-          >
-            ${formatCurrencyBalance(balance)}
-          </Text>
-        </div>
-        <div className={styles.usdcInfo}>
-          <Text color='staticWhite'>{messages.earn}</Text>
-          <HarmonyPlainButton
-            // TODO: wire up learn more link https://linear.app/audius/issue/PAY-1762/wire-up-learn-more-link
-            onClick={() => {}}
-            iconLeft={IconQuestionCircle}
-            variant={HarmonyPlainButtonType.INVERTED}
-            text={messages.learnMore}
-          />
-        </div>
-      </div>
-      <div className={styles.withdrawContainer}>
-        <HarmonyButton
-          variant={HarmonyButtonType.SECONDARY}
-          text={messages.withdraw}
-          // TODO: update leftIcon and wire up withdraw modal https://linear.app/audius/issue/PAY-1754/usdc-withdrawal-flow-ui
-          iconLeft={() => <Icon icon={IconNote} size='medium' />}
-          onClick={() => {}}
-        />
-        <PopupMenu
-          transformOrigin={{ horizontal: 'center', vertical: 'top' }}
-          anchorOrigin={{ horizontal: 'center', vertical: 'bottom' }}
-          items={menuItems}
-          renderTrigger={(anchorRef, triggerPopup) => (
-            <HarmonyButton
-              ref={anchorRef}
-              variant={HarmonyButtonType.SECONDARY}
-              iconLeft={IconKebabHorizontal}
-              onClick={triggerPopup}
-            />
-          )}
-        />
-      </div>
-    </div>
-  )
+  thisYear: 'This Year'
 }
 
 type ArtistDashboardPageProps = ReturnType<typeof mapDispatchToProps> &
@@ -503,7 +229,7 @@ export class ArtistDashboardPage extends Component<
         ) : (
           <>
             {this.renderProfileSection()}
-            {isUSDCEnabled ? <USDCSection account={account} /> : null}
+            {isUSDCEnabled ? <USDCTile balance={0} /> : null}
             {this.renderCreatorContent()}
           </>
         )}
